pyproject.toml
[project]
name = "trp1-ledger"
version = "0.1.0"
description = "TRP1 Week 5 - Agentic Event Store & Audit Infrastructure (interim)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.29.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
]
